test(canon): split reachability budget regressions
